+++ TFL_API_Framework/TFL_API_TestFramework/TFL_API_TestFramework.csproj
@@ -58,7 +58,7 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="WhenRouteIsCalledWithValidLineName.cs" />
+    <Compile Include="SingleRouteVictoriaLineResponseServiceTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
